Adicionado o ResponseEntity

+++ src/main/java/com/example/API/controller/UsuarioController/UsuarioController.java
@@ -22,30 +22,36 @@ public class UsuarioController {
     public ResponseEntity<Usuario> getUsuario(@PathVariable Long id) {
         Usuario usuario = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
-        return ResponseEntity.ok(usuario);
+        return ResponseEntity.ok(usuario); // 200 OK com a lista
     }
 
     @PostMapping
-    public void salvarUsuario(@RequestBody UsuarioRequestDTO usuarioData){
+    public ResponseEntity<Void> salvarUsuario(@RequestBody UsuarioRequestDTO usuarioData){
         Usuario usuario = new Usuario(usuarioData);
         repository.save(usuario);
-        return;
+        return ResponseEntity.status(HttpStatus.CREATED).build(); // 201 Created
     }
 
     @PutMapping("/{id}")
-    public void alterarUsuario(@PathVariable Long id, @RequestBody UsuarioRequestDTO usuarioData){
+    public ResponseEntity<Void> alterarUsuario(@PathVariable Long id, @RequestBody UsuarioRequestDTO usuarioData){
         Usuario usuario = repository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Usuario não encontrado"));
         usuarioData.atualizarEntidade(usuario);
         repository.save(usuario);
+        return ResponseEntity.noContent().build() ; // 200 OK Sem Body
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<UsuarioController.MensagemResponse> DeletarProdutos(@PathVariable Long id){
+        if (!repository.existsById(id)){
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND) // 404
+                    .body(new MensagemResponse("Produto não encontrado"));
+        }
         repository.deleteById(id);
         return ResponseEntity
-                .status(HttpStatus.OK)
-                .body(new MensagemResponse("Produto deletado com sucesso"));
+                .status(HttpStatus.OK) // 200 OK sem Body
+                .build();
     }
 
 }
